简单调整结构
