+++ .github/workflows/main.yml
@@ -79,6 +79,6 @@ jobs:
         run: echo "TAG_NAME=${GITHUB_REF#refs/tags/}" >> $GITHUB_ENV
 
       - name: Upload Files
-        run: gh release upload --clobber ${{ env.TAG_NAME }} plugin_test.py
+        run: gh release upload --clobber ${{ env.TAG_NAME }} src/plugins/store/plugin_test.py
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
