+++ src/plugins/store/constants.py
@@ -1,0 +1,2 @@
+PLUGIN_CONFIGS_URL = "https://raw.githubusercontent.com/he0119/nonebot-store-test/main/inputs/configs.json"
+RESULTS_URL = "https://raw.githubusercontent.com/he0119/nonebot-store-test/main/results/results.json"
