+++ src/plugins/store/models.py
@@ -1,0 +1,11 @@
+from typing import TypedDict
+
+
+class Metadata(TypedDict):
+    """插件元数据"""
+
+    name: str
+    description: str
+    homepage: str
+    type: str
+    supported_adapters: list[str]
