+++ src/plugins/store/utils.py
@@ -1,0 +1,9 @@
+import re
+
+
+def strip_ansi(text: str | None) -> str:
+    """去除 ANSI 转义字符"""
+    if not text:
+        return ""
+    ansi_escape = re.compile(r"\x1B(?:[@-Z\\-_]|\[[0-?]*[ -/]*[@-~])")
+    return ansi_escape.sub("", text)
